Update hashlib adapter
